+++ .gitmodules
@@ -1,5 +1,5 @@
-[submodule "rosflight/include/mavlink/v1.0"]
-	path = rosflight/include/rosflight/mavlink/v1.0
+[submodule "rosflight_io/include/rosflight_io/mavlink/v1.0"]
+	path = rosflight_io/include/rosflight_io/mavlink/v1.0
 	url = https://github.com/rosflight/mavlink_c_library.git
 [submodule "rosflight_firmware/firmware"]
 	path = rosflight_firmware/firmware
