+++ rosflight_io/include/rosflight_io/mavrosflight/mavlink_serial.hpp
@@ -30,50 +30,71 @@
  */
 
 /**
- * \file mavrosflight.h
+ * \file mavlink_serial.h
  * \author Daniel Koch <daniel.koch@byu.edu>
  */
 
-#ifndef MAVROSFLIGHT_MAVROSFLIGHT_H
-#define MAVROSFLIGHT_MAVROSFLIGHT_H
+#ifndef MAVROSFLIGHT_MAVLINK_SERIAL_H
+#define MAVROSFLIGHT_MAVLINK_SERIAL_H
 
-#include <rosflight/mavrosflight/mavlink_bridge.hpp>
-#include <rosflight/mavrosflight/mavlink_comm.hpp>
-#include <rosflight/mavrosflight/param_manager.hpp>
-#include <rosflight/mavrosflight/time_manager.hpp>
+#include <rosflight_io/mavrosflight/mavlink_comm.hpp>
 
-#include <rosflight/mavrosflight/mavlink_listener_interface.hpp>
-#include <rosflight/mavrosflight/param_listener_interface.hpp>
-
-#include <rclcpp/rclcpp.hpp>
-
+#include <boost/asio.hpp>
 #include <boost/function.hpp>
 
-#include <cstdint>
 #include <string>
 
 namespace mavrosflight
 {
-class MavROSflight
+class MavlinkSerial : public MavlinkComm
 {
 public:
   /**
    * \brief Instantiates the class and begins communication on the specified serial port
-   * \param mavlink_comm Reference to a MavlinkComm object (serial or UDP)
+   * \param port Name of the serial port (e.g. "/dev/ttyUSB0")
    * \param baud_rate Serial communication baud rate
    */
-  MavROSflight(MavlinkComm & mavlink_comm, rclcpp::Node * node);
+  MavlinkSerial(std::string port, int baud_rate);
 
   /**
    * \brief Stops communication and closes the serial port before the object is destroyed
    */
-  ~MavROSflight();
+  ~MavlinkSerial() override;
 
-  MavlinkComm & comm;
-  ParamManager param;
-  TimeManager time;
+private:
+  //===========================================================================
+  // methods
+  //===========================================================================
+
+  bool is_open() override;
+  void do_open() override;
+  void do_close() override;
+
+  /**
+   * \brief Initiate an asynchronous read operation
+   */
+  void do_async_read(
+    const boost::asio::mutable_buffers_1 & buffer,
+    boost::function<void(const boost::system::error_code &, size_t)> handler) override;
+
+  /**
+   * \brief Initialize an asynchronous write operation
+   * \param check_write_state If true, only start another write operation if a write sequence is not already running
+   */
+  void do_async_write(
+    const boost::asio::const_buffers_1 & buffer,
+    boost::function<void(const boost::system::error_code &, size_t)> handler) override;
+
+  //===========================================================================
+  // member variables
+  //===========================================================================
+
+  boost::asio::serial_port serial_port_; //!< boost serial port object
+
+  std::string port_;
+  int baud_rate_;
 };
 
 } // namespace mavrosflight
 
-#endif // MAVROSFLIGHT_MAVROSFLIGHT_H
+#endif // MAVROSFLIGHT_MAVLINK_SERIAL_H
