+++ rosflight_io/include/rosflight_io/mavrosflight/serial_exception.hpp
