+++ rosflight_io/src/mag_cal_node.cpp
@@ -1,5 +1,8 @@
 /*
+ * Software License Agreement (BSD-3 License)
+ *
  * Copyright (c) 2017 Daniel Koch and James Jackson, BYU MAGICC Lab.
+ * Copyright (c) 2023 Brandon Sutherland, AeroVironment Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,76 +33,20 @@
  */
 
 /**
- * \file param.h
- * \author Daniel Koch <daniel.koch@byu.edu>
+ * \file mag_cal_node.cpp
+ * \author Devon Morris <devonmorris1992@gmail.com>
  */
 
-#ifndef MAVROSFLIGHT_PARAM_H
-#define MAVROSFLIGHT_PARAM_H
+#include <rclcpp/rclcpp.hpp>
+#include <rosflight_io/mag_cal.hpp>
 
-#include <rosflight/mavrosflight/mavlink_bridge.hpp>
-#include <rosflight/mavrosflight/mavlink_serial.hpp>
-
-#include <yaml-cpp/yaml.h>
-
-namespace mavrosflight
+int main(int argc, char ** argv)
 {
-class Param
-{
-public:
-  Param();
-  explicit Param(mavlink_param_value_t msg);
-  Param(std::string name, int index, MAV_PARAM_TYPE type, float raw_value);
+  rclcpp::init(argc, argv);
 
-  std::string getName() const;
-  int getIndex() const;
-  MAV_PARAM_TYPE getType() const;
-  double getValue() const;
+  auto calibrate_node = std::make_shared<rosflight_io::CalibrateMag>();
+  calibrate_node->run();
 
-  void requestSet(double value, mavlink_message_t * msg);
-  bool handleUpdate(const mavlink_param_value_t & msg);
-
-private:
-  void init(std::string name, int index, MAV_PARAM_TYPE type, float raw_value);
-
-  void setFromRawValue(float raw_value);
-  float getRawValue();
-  float getRawValue(double value);
-  double getCastValue(double value);
-
-  template<typename T>
-  double fromRawValue(float value)
-  {
-    T t_value;
-    std::memcpy(&t_value, &value, sizeof(T));
-    return static_cast<double>(t_value);
-  }
-
-  template<typename T>
-  float toRawValue(double value)
-  {
-    T t_value = static_cast<T>(value);
-    float result = 0.0f;
-    std::memcpy(&result, &t_value, sizeof(T));
-    return result;
-  }
-
-  template<typename T>
-  double toCastValue(double value)
-  {
-    return static_cast<double>(static_cast<T>(value));
-  }
-
-  std::string name_;
-  int index_;
-  MAV_PARAM_TYPE type_;
-  double value_;
-
-  bool set_in_progress_;
-  double new_value_;
-  float expected_raw_value_;
-};
-
-} // namespace mavrosflight
-
-#endif // MAVROSFLIGHT_PARAM_H
+  rclcpp::shutdown();
+  return 0;
+}
