+++ rosflight_io/src/mavrosflight/mavlink_serial.cpp
@@ -30,15 +30,53 @@
  */
 
 /**
- * \file mavlink_bridge.h
+ * \file mavlink_serial.cpp
  * \author Daniel Koch <daniel.koch@byu.edu>
- *
- * MAVLink adapter header
  */
 
-#ifndef MAVROSFLIGHT_MAVLINK_BRIDGE_H
-#define MAVROSFLIGHT_MAVLINK_BRIDGE_H
+#include <rosflight_io/mavrosflight/mavlink_serial.hpp>
+#include <rosflight_io/mavrosflight/serial_exception.hpp>
 
-#include <rosflight/mavlink/v1.0/rosflight/mavlink.h>
+namespace mavrosflight
+{
+using boost::asio::serial_port_base;
 
-#endif // MAVROSFLIGHT_MAVLINK_BRIDGE_H
+MavlinkSerial::MavlinkSerial(std::string port, int baud_rate)
+    : MavlinkComm(), serial_port_(io_service_), port_(std::move(port)), baud_rate_(baud_rate)
+{}
+
+MavlinkSerial::~MavlinkSerial() { MavlinkSerial::do_close(); }
+
+bool MavlinkSerial::is_open() { return serial_port_.is_open(); }
+
+void MavlinkSerial::do_open()
+{
+  try {
+    serial_port_.open(port_);
+    serial_port_.set_option(serial_port_base::baud_rate(baud_rate_));
+    serial_port_.set_option(serial_port_base::character_size(8));
+    serial_port_.set_option(serial_port_base::parity(serial_port_base::parity::none));
+    serial_port_.set_option(serial_port_base::stop_bits(serial_port_base::stop_bits::one));
+    serial_port_.set_option(serial_port_base::flow_control(serial_port_base::flow_control::none));
+  } catch (const boost::system::system_error & e) {
+    throw SerialException(e);
+  }
+}
+
+void MavlinkSerial::do_close() { serial_port_.close(); }
+
+void MavlinkSerial::do_async_read(
+  const boost::asio::mutable_buffers_1 & buffer,
+  boost::function<void(const boost::system::error_code &, size_t)> handler)
+{
+  serial_port_.async_read_some(buffer, handler);
+}
+
+void MavlinkSerial::do_async_write(
+  const boost::asio::const_buffers_1 & buffer,
+  boost::function<void(const boost::system::error_code &, size_t)> handler)
+{
+  serial_port_.async_write_some(buffer, handler);
+}
+
+} // namespace mavrosflight
