+++ rosflight_msgs/package.xml
@@ -2,7 +2,7 @@
 <package format="3">
   <name>rosflight_msgs</name>
   <version>2.0.0</version>
-  <description>Message and service definitions for the ROSflight2 ROS 2 stack</description>
+  <description>Message and service definitions for the ROSflight ROS stack</description>
 
   <maintainer email="daniel.p.koch@gmail.com">Daniel Koch</maintainer>
